% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spinebil.R
\name{basisMatrix}
\alias{basisMatrix}
\title{Generate 2-d basis in directions i, j in n dimensions (i,j <= n)}
\usage{
basisMatrix(i, j, n)
}
\arguments{
\item{i}{first basis direction}

\item{j}{second basis direction}

\item{n}{number of dimensions}
}
\value{
basis matrix
}
\description{
Generate 2-d basis in directions i, j in n dimensions (i,j <= n)
}
