% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-spinebil.R
\name{basisVector}
\alias{basisVector}
\title{Generate basis vector in direction i in n dimensions (i <= n)}
\usage{
basisVector(i, n)
}
\arguments{
\item{i}{selected direction}

\item{n}{number of dimensions}
}
\value{
basis vector
}
\description{
Generate basis vector in direction i in n dimensions (i <= n)
}
