% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{centerAll}
\alias{centerAll}
\title{Combined centering for projected data and cube points}
\usage{
centerAll(dPoints, cPoints)
}
\arguments{
\item{dPoints}{Projected data points (matrix)}

\item{cPoints}{Projected hypercube points (matrix)}
}
\value{
Named list containint the centered data points ("data")
    and cube points ("cube")
}
\description{
For each view we center all points to be drawn.
}
