% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{colorList}
\alias{colorList}
\title{Mapping grouping to color.}
\usage{
colorList(gr)
}
\arguments{
\item{gr}{Vector containing group assignment for each entry.}
}
\value{
Named list containing assigned colors ("color")
    and list of colors ("col")
}
\description{
Mapping grouping to color.
}
\keyword{internal}
