% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{coverageDispMargin}
\alias{coverageDispMargin}
\title{Formatted margins for the coverage display in plotly.}
\format{An object of class \code{list} of length 4.}
\usage{
coverageDispMargin
}
\description{
Formatted margins for the coverage display in plotly.
}
\keyword{internal}
