% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{cubeLineStyle}
\alias{cubeLineStyle}
\title{Formatted line style for drawing the cube outline for plotly.}
\format{An object of class \code{list} of length 2.}
\usage{
cubeLineStyle
}
\description{
Formatted line style for drawing the cube outline for plotly.
}
\keyword{internal}
