% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{cubePoints}
\alias{cubePoints}
\title{Generate vertices on hypercube enclosing all data points.}
\usage{
cubePoints(n, d)
}
\arguments{
\item{n}{Number of parameters in the input data.}

\item{d}{Input data (numeric matrix).}
}
\value{
Data frame containing all vertices of the hypercube
}
\description{
Generate vertices on hypercube enclosing all data points.
}
