% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{customLegend}
\alias{customLegend}
\title{Build a legend of colored label text.}
\usage{
customLegend(labs, col, halfRange)
}
\arguments{
\item{labs}{Text vector containing all labels.}

\item{col}{Color vector containing corresponding colors.}

\item{halfRange}{Half range parameter used in the plot (for positioning)}
}
\value{
Vector of list to use as annotation in a plotly layout call for
    generating the custom legend.
}
\description{
Build a legend of colored label text.
}
\keyword{internal}
