% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{dcorIndex}
\alias{dcorIndex}
\title{Distance correlation based tour index}
\usage{
dcorIndex()
}
\value{
Function taking matrix input data (projected data),
    returning value of dcor2d metric.
}
\description{
Distance correlation based tour index
}
