% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{formatProj}
\alias{formatProj}
\title{Formatting projection matrix for screen printout.}
\usage{
formatProj(proj, params, idx)
}
\arguments{
\item{proj}{Projection matrix to be formatted}

\item{params}{Parameter names}

\item{idx}{Index value of this projection in the tour sequence}
}
\value{
Formatted text for screen printout.
}
\description{
Formatting projection matrix for screen printout.
}
