% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{fullTourPCA}
\alias{fullTourPCA}
\title{PCA over the projection vectors in the tour sequence.}
\usage{
fullTourPCA(fullTour, n)
}
\arguments{
\item{fullTour}{List of projections in interpolated tour path.}

\item{n}{Number of parameters.}
}
\value{
prcomp output over all vectors in the path.
}
\description{
PCA over the projection vectors in the tour sequence.
}
