% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{galah}
\alias{galah}
\title{Scatter points containing a galah.}
\format{A data frame with 553 rows and 7 variables (6 parameters and color)}
\usage{
galah
}
\description{
Try plotting x vs y or x1 vs y1!
}
\keyword{datasets}
