% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/galahr-package.R
\docType{package}
\name{galahr}
\alias{galahr}
\alias{galahr-package}
\title{galahr: A GUI to assist learning about high dimensions in R}
\description{
A graphical interface to tourr using Shiny and
plotly to explore high dimensional distributions through
dynamic 2-d projections.
}
\seealso{
Launch the app by calling \code{\link[=launchApp]{launchApp()}}
}
