% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\name{getGuidedTour}
\alias{getGuidedTour}
\title{Taking guided tour index name and returning corresponding guided tour}
\usage{
getGuidedTour(indexName, grId = NA)
}
\arguments{
\item{indexName}{Index name from \code{\link{guidedTourOptions}}}

\item{grId}{Class to be used when using \code{\link{groupedIndex}}}
}
\value{
guided_tour with corresponding index function
}
\description{
Taking guided tour index name and returning corresponding guided tour
}
