% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\name{getPathIndex}
\alias{getPathIndex}
\title{Taking interpolated tour path and index name and returning path index}
\usage{
getPathIndex(fullTour, indexName, grId = NA)
}
\arguments{
\item{fullTour}{Interpolated tour path}

\item{indexName}{Index name from \code{\link{guidedTourOptions}}}

\item{grId}{Class to be used when using \code{\link{groupedIndex}}}
}
\value{
Index values over the tour history
}
\description{
Taking interpolated tour path and index name and returning path index
}
