% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{getSmallMarker}
\alias{getSmallMarker}
\title{Generating smaller marker styles for plotly traces.}
\usage{
getSmallMarker(col, a = NULL)
}
\arguments{
\item{col}{Marker color as string (black, red).}

\item{a}{Transperancy (default is not transperant.)}
}
\value{
List formatted as required for plotly marker style.
}
\description{
Generating smaller marker styles for plotly traces.
}
\keyword{internal}
