% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\docType{data}
\name{groupedIndex}
\alias{groupedIndex}
\title{List of names for all available guided tour index functions with grouping.}
\format{An object of class \code{character} of length 2.}
\usage{
groupedIndex
}
\description{
List of names for all available guided tour index functions with grouping.
}
\keyword{internal}
