% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexFunctions.R
\docType{data}
\name{guidedTourOptions}
\alias{guidedTourOptions}
\title{List of names for all available guided tour index functions.}
\format{An object of class \code{character} of length 10.}
\usage{
guidedTourOptions
}
\description{
List of names for all available guided tour index functions.
}
\keyword{internal}
