% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GW170817}
\alias{GW170817}
\title{Posterior samples fitting a gravitational wave observation.}
\format{A tibble with 2396 rows and 6 variables}
\source{
\url{https://dcc.ligo.org/public/0152/P1800115/005/Parametrized-EoS_maxmass_posterior_samples.dat}
}
\usage{
GW170817
}
\description{
A dataset containing posterior samples describing neutron star
properties measured by LIGO in the gravitational wave observation
GW170817.
}
\references{
\url{https://doi.org/10.1103/PhysRevLett.119.161101}
}
\keyword{datasets}
