% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{hoverText}
\alias{hoverText}
\title{Generate hover text for each data point.}
\usage{
hoverText(d, p)
}
\arguments{
\item{d}{Data frame containing input data}

\item{p}{Parameters to include in hover text.}
}
\value{
Data frame with new column paramT containing the hover text.
}
\description{
Generate hover text for each data point.
}
\keyword{internal}
