% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{initializeReactive}
\alias{initializeReactive}
\title{Initialising the reactive values used in the app.}
\usage{
initializeReactive(paramDF)
}
\arguments{
\item{paramDF}{Dataframe containing the input data.}
}
\value{
Reactive value container.
}
\description{
Initialising the reactive values used in the app.
}
\keyword{internal}
