% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{isEven}
\alias{isEven}
\title{Simple helper to test if number is even.}
\usage{
isEven(x)
}
\arguments{
\item{x}{Numeric input to be tested.}
}
\value{
Logical vector
}
\description{
Simple helper to test if number is even.
}
\keyword{internal}
