% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for exploring high dimensional data}
\usage{
launchApp(paramDF = NULL)
}
\arguments{
\item{paramDF}{input data to load the app with}
}
\description{
If no input data is specified the app will launch
using the [tsfeatureData] example data.
}
\examples{
\dontrun{
launchApp()
launchApp(paramDF = GW170817)
}
}
