% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{mineIndex}
\alias{mineIndex}
\title{MINE based tour index}
\usage{
mineIndex(mineIndex)
}
\arguments{
\item{mineIndex}{MINE index name to be used}
}
\value{
Function taking matrix input data (projected data),
    returning value of selected MINE metric.
}
\description{
Available indexes are MIC, TIC
}
