% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{noAxis}
\alias{noAxis}
\title{Formatted empty axis style for plotly.}
\format{An object of class \code{list} of length 5.}
\usage{
noAxis
}
\description{
Formatted empty axis style for plotly.
}
\keyword{internal}
