% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{noAxisRange}
\alias{noAxisRange}
\title{Formatted empty axis style with fixed range used for timeline plotting.}
\usage{
noAxisRange(axmin, axmax)
}
\arguments{
\item{axmin}{Lower limit of axis}

\item{axmax}{Upper limit of axis}
}
\value{
List formatted as required for plotly axis style.
}
\description{
Formatted empty axis style with fixed range used for timeline plotting.
}
\keyword{internal}
