% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_1d}
\alias{plotly_1d}
\title{Generating the plotly 1-d parameter display.}
\usage{
plotly_1d(d1, d2, markerD1 = NULL)
}
\arguments{
\item{d1}{Data points in currently selected sample}

\item{d2}{Data points outside the currently selected sample}

\item{markerD1}{Custom marker (list) for d1, if NULL use default markers.}
}
\value{
List of plotly visualisations
}
\description{
This function generates a list of 1-d plotly visualisations
of the parameter values found in the sample. Data points outside
the sample are plotted as invisible markers ensuring that we span the
full range for each parameter. Custom markers can be passed in for
data points inside the selected sample. This plot is used in the app
for linked brushing, highlighting points in the tour display based on
selection on parameter values.
}
