% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotly_tour_grouped}
\alias{plotly_tour_grouped}
\title{Generating the plotly tour display for grouped data.}
\usage{
plotly_tour_grouped(scatterData, cubeData, hoverData, halfRange, gr)
}
\arguments{
\item{scatterData}{Projected data points (matrix)}

\item{cubeData}{Projected hypercube points (matrix)}

\item{hoverData}{Data frame containing hover text in "paramT" column}

\item{halfRange}{Half range for fixing axis lenght}

\item{gr}{Vector containing group assignment for each data entry}
}
\value{
plotly visualisation
}
\description{
Generating the plotly tour display for grouped data.
}
