% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotlyAxesF}
\alias{plotlyAxesF}
\title{Generating the plotly axes display.}
\usage{
plotlyAxesF(xVec, yVec, paramList)
}
\arguments{
\item{xVec}{Vector definig x direction of the projection}

\item{yVec}{Vector definig y direction of the projection}

\item{paramList}{Vector of parameter names}
}
\value{
Plotly visualisation of axes
}
\description{
Generating the plotly axes display.
}
