% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\docType{data}
\name{plotlyMargin}
\alias{plotlyMargin}
\title{Formatted margins for the timeline in plotly.}
\format{An object of class \code{list} of length 4.}
\usage{
plotlyMargin
}
\description{
Formatted margins for the timeline in plotly.
}
\keyword{internal}
