% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotlyTourF}
\alias{plotlyTourF}
\title{Generating the plotly tour display.}
\usage{
plotlyTourF(scatterData, cubeData, hoverData, halfRange, red = FALSE)
}
\arguments{
\item{scatterData}{Projected data points (matrix)}

\item{cubeData}{Projected hypercube points (matrix)}

\item{hoverData}{Data frame containing hover text in "paramT" column}

\item{halfRange}{Half range for fixing axis lenght}

\item{red}{Logical, if TRUE use red markers for the scatter points}
}
\value{
Plotly visualisation
}
\description{
Generating the plotly tour display.
}
