% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUpdates.R
\name{readInput}
\alias{readInput}
\title{Read input data and update UI accordingly.}
\usage{
readInput(file1, rv, output, session)
}
\arguments{
\item{file1}{fileInput from shiny}

\item{rv}{Reactive value container}

\item{output}{shiny output container}

\item{session}{shiny session}
}
\description{
Read input data and update UI accordingly.
}
\keyword{internal}
