% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{scags}
\alias{scags}
\title{Scagnostics based tour index}
\usage{
scags(scagMetricIndex)
}
\arguments{
\item{scagMetricIndex}{Scagnostics index name to be used}
}
\value{
Function taking matrix input data (projected data),
returning value of selected scagnostics metric.
}
\description{
Available scagnostics indexes are: Skinny, Striated, Convex, Clumpy
}
