% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{splineIndex}
\alias{splineIndex}
\title{Spline based tour index}
\usage{
splineIndex()
}
\value{
Function taking matrix input data (projected data),
    returning value of splines2d metric.
}
\description{
Spline based tour index
}
