% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputUpdates.R
\name{splitInput}
\alias{splitInput}
\title{Split input data frame into numeric and factor columns.}
\usage{
splitInput(inData, rv)
}
\arguments{
\item{inData}{Input dataframe}

\item{rv}{Reactive value container.}
}
\description{
Split input data frame into numeric and factor columns.
}
\keyword{internal}
