% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tabInput}
\alias{tabInput}
\title{Generating layout for the input tab panel.}
\usage{
tabInput(params)
}
\arguments{
\item{params}{The input parameters for parameter selection checkbox input.}
}
\value{
The complete shiny tabPanel.
}
\description{
Generating layout for the input tab panel.
}
\keyword{internal}
