% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{tabResults}
\alias{tabResults}
\title{Generating layout for the results tab panel.}
\usage{
tabResults(npoint)
}
\arguments{
\item{npoint}{The number of points in the input data set.}
}
\value{
The complete shiny tabPanel.
}
\description{
Generating layout for the results tab panel.
}
\keyword{internal}
