% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{timelineAxis}
\alias{timelineAxis}
\title{Formatted axis style with fixed range (0, xMax+1) and breaks.}
\usage{
timelineAxis(xMax, breaks)
}
\arguments{
\item{xMax}{Upper limit for axis range}

\item{breaks}{Values for axis ticks}
}
\value{
List formatted as required for plotly axis style.
}
\description{
Formatted axis style with fixed range (0, xMax+1) and breaks.
}
\keyword{internal}
