% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{tourAxis}
\alias{tourAxis}
\title{Formatted empty axis style with fixed range used for tour plotting.}
\usage{
tourAxis(halfRange)
}
\arguments{
\item{halfRange}{Axis range will be between -/+ halfRange.}
}
\value{
List formatted as required for plotly axis style.
}
\description{
Formatted empty axis style with fixed range used for tour plotting.
}
\keyword{internal}
