% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsfeatureData}
\alias{tsfeatureData}
\title{Features calculated on different types of time series.}
\format{A tibble with 1082 rows and 14 variables (13 features and type)}
\source{
\url{https://www.comp-engine.org}
}
\usage{
tsfeatureData
}
\description{
A dataset containing the features calculated with `tsfeatures`
on three groups of time series: bird songs, air temperature and
music.
}
\references{
\url{https://doi.org/10.1098/rsif.2013.0048}
}
\keyword{datasets}
