% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Generating the shiny ui.}
\usage{
ui(params, npoint)
}
\arguments{
\item{params}{The input parameters for parameter selection checkbox input.}

\item{npoint}{The number of points in the input data set.}
}
\value{
The complete shiny fluidPage UI.
}
\description{
Generating the shiny ui.
}
\keyword{internal}
