% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{updatePlots}
\alias{updatePlots}
\title{Update all plots to current projection.}
\usage{
updatePlots(rv, session, input, output)
}
\arguments{
\item{rv}{Reactive value container}

\item{session}{shiny session}

\item{input}{shiny input container}

\item{output}{shiny output container}
}
\description{
Update all plots to current projection.
}
\keyword{internal}
