% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{updateReactiveData}
\alias{updateReactiveData}
\title{Updating relevant reactive values.}
\usage{
updateReactiveData(rv)
}
\arguments{
\item{rv}{Reactive value container}
}
\description{
Updating relevant reactive values.
}
\keyword{internal}
