% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{coveragePlot}
\alias{coveragePlot}
\title{Generating the coverage display.}
\usage{
coveragePlot(pcaRes, n, i)
}
\arguments{
\item{pcaRes}{Results from \code{\link[=fullTourPCA]{fullTourPCA()}}}

\item{n}{Number of input parameters}

\item{i}{Index of current projection}
}
\value{
Visualisation of coverage display.
}
\description{
Generating the coverage display.
}
