% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchApp}
\alias{launchApp}
\title{Shiny app for exploring high dimensional data}
\usage{
launchApp(paramDF = NULL)
}
\arguments{
\item{paramDF}{input data to load the app with. By default load data from FIXME SOME EXAMPLE}
}
\description{
Shiny app for exploring high dimensional data
}
\examples{
\dontrun{
launch_app()
FIXME example with alternative input data
}
}
